Relocate spatial visitor into separate file
